move IndexKey and IndexKeyItem to typing module, add to docs, resolve circular deps
