+++ docs/src/reference/temporian/internals/typing/IndexKey.md
@@ -1,0 +1,1 @@
+::: temporian.core.typing.IndexKey
