+++ docs/src/reference/temporian/internals/typing/IndexKeyItem.md
@@ -1,0 +1,1 @@
+::: temporian.core.typing.IndexKeyItem
