+++ temporian/core/operators/base.py
@@ -18,9 +18,10 @@ from __future__ import annotations
 from abc import ABC
 from copy import deepcopy
 from typing import Dict, List, Tuple, Union, Any
-from temporian.core.data.dtype import DType, IndexKey
+from temporian.core.data.dtype import DType
 
 from temporian.core.data.node import EventSetNode
+from temporian.core.typing import IndexKey
 from temporian.proto import core_pb2 as pb
